requirements.txt
# Document Analyzer - Improved Version
# Python 3.10+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
psycopg2-pool==1.1
SQLAlchemy==2.0.25  # Optional for future ORM migration

# LLM & AI
openai==1.10.0
anthropic==0.18.1
langchain==0.1.6
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-core==0.1.21
langchain-community==0.0.19
langsmith==0.0.87

# Vector Store
pinecone-client==3.0.2
sentence-transformers==2.3.1

# AWS
boto3==1.34.34
botocore==1.34.34

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.2.0

# LlamaParse (optional)
llama-parse==0.4.0

# Structured Logging
structlog==24.1.0

# Redis for rate limiting and caching
redis==5.0.1

# Prometheus metrics
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
tenacity==8.2.3
requests==2.31.0

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality (optional)
black==24.1.1
ruff==0.1.14
mypy==1.8.0

